Gracefully stop when no precondition holds
Imagine that an (expected) error is thrown during the tests and the tested
system is now in a state where it can no longer process new commands. In that
case we should be able to gracefully stop the test in progress.

Simply stop the test when no precondition holds instead of reporting a
deadlock error. Labelling can be used to detect such deadlocks.

Signed-off-by: Thomas Winant <thomas@well-typed.com>