Allow the generation of commands to stop
Imagine that an (expected) error is thrown during the tests and the tested
system is now in a state where it can no longer process new commands. In that
case we should be able to gracefully stop the test in progress.

We achieve this by letting the generator return `Nothing` to indicate that no
more commands should be generated.

Signed-off-by: Thomas Winant <thomas@well-typed.com>