High-level interface to the state machine API
This captures the patterns described in
http://www.well-typed.com/blog/2019/01/qsm-in-depth/ as a proper Haskell
abstraction. The basic idea is that we have a concept of commands and responses
(just like in the lower level API), along with two interpretors, one for the
system under test and one for the model. We then compare that they can run in
lockstep, comparing responses at every step, insisting that they must be equal.

Of course requiring that responses must be really _equal_ is too strong: the
system under test might return a real file handle for example, where the model
will return a mock file handle. So we compare responses up to a mapping from
real handles to mock handles, which is maintained by the Lockstep
infrastructure.

It comes in two variants: one n-ary variant supporting an arbitrary number of
different types of handles, and one where there is precisely one. The n-ary
version is needed if testing for example a system that might return both DB
handles and file handles, but in most cases the simple one suffices and is much
easier to use. The simple API is defined in terms of the n-ary one (which
requires some more sophisticated types).

The IORefs test is a nice example that demonstrates quite how succinctly we can
write model tests with these abstractions.

This is joint work with Alfredo Di Napoli.

Signed-off-by: Edsko de Vries <edsko@well-typed.com>