Avoid unnecessary use of `run`
The top-level of a QSM test typically looks something like this:

```haskell
forAllCommands ... $ \cmds -> QC.monadicIO $ do
  (hist, prop) <- .... runCommands ...
  prettyCommands ...
```

The problem is that `runCommands` lives in `PropertyM`. This makes it
impossible to do any kind of bracketing. Fortunately, it turns out that this
use of `PropertyM` is actually not needed at all. In this commit, I add
`runCommands'`, which lives in `m`. To avoid breaking backwards compatibility,
`runCommands` still lives in `PropertyM` and is a combination of `run` and
`runCommands'`.

This will allow us to do something like:

```haskell
forAllCommands ... $ \cmds -> QC.monadicIO $ do
  (hist, prop) <- .... run (bracket ... $ \r -> runCommands' .. r .. ) ...
  prettyCommands ...
```

Co-authored-by: Thomas Winant <thomas@well-typed.com>
Signed-off-by: Edsko de Vries <edsko@well-typed.com>